a bunch of code style/terminology changes, no functional changes
details:
	change all remaining instances of "type * var" to "type *var"
	change all instances of "height" to "lines" and "width" to "cols"
	change the order of some function args to lines, cols instead of width, height
		...but don't change ltm_set_window_dimensions yet, because that would be a
		functionality change
	use more appropriate/correct types in some places
	add missing semicolons to the end of some macro calls
	change some sizeofs so that they get the type from a variable instead of
		hard-coding the type
	some incredibly minor wording changes in comments

+++ src/cursor.h
@@ -9,8 +9,8 @@
 #define X 0
 #define Y 1
 
-extern void cursor_rel_move(int, char, uint);
-extern void cursor_abs_move(int, char, uint);
+extern void cursor_rel_move(int, char, ushort);
+extern void cursor_abs_move(int, char, ushort);
 
 extern void cursor_down(int);
 extern void cursor_wrap(int);
