+++ src/window.h
@@ -1,7 +1,7 @@
 #ifndef WINDOW_H
 #define WINDOW_H
 
-extern int ltm_set_window_dimensions(int, uint, uint);
+extern int ltm_set_window_dimensions(int, ushort, ushort);
 extern int tcsetwinsz(int);
 extern int scroll_screen(int);
